requirements.txt
alembic==1.8.0
anyio==3.6.1
apache-airflow==2.3.2
apache-airflow-providers-ftp==2.1.2
apache-airflow-providers-http==2.1.2
apache-airflow-providers-imap==2.2.3
apache-airflow-providers-sqlite==2.1.3
apispec==3.3.2
argcomplete==2.0.0
attrs==20.3.0
Babel==2.10.1
blinker==1.4
Brotli==1.0.9
cachelib==0.7.0
cachetools==5.1.0
cattrs==1.10.0
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.5
click==7.1.2
clickclick==20.10.2
colorama==0.4.4
colorlog==4.8.0
commonmark==0.9.1
connexion==2.13.1
cron-descriptor==1.2.24
croniter==1.3.5
cryptography==37.0.2
cycler==0.11.0
dash==2.5.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
defusedxml==0.7.1
Deprecated==1.2.13
dill==0.3.5.1
dnspython==2.2.1
docutils==0.18.1
email-validator==1.2.1
Flask==1.1.4
Flask-AppBuilder==3.4.5
Flask-Assets==2.0
Flask-Babel==2.0.0
Flask-Caching==1.11.1
Flask-Compress==1.12
Flask-DotEnv==0.1.2
Flask-JWT-Extended==3.25.1
Flask-Login==0.4.1
Flask-OpenID==1.3.0
Flask-Session==0.4.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.14.3
fonttools==4.28.3
future==0.18.2
google-api-core==2.8.1
google-api-python-client==2.49.0
google-auth==2.6.6
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.1
googleapis-common-protos==1.56.2
graphviz==0.20
greenlet==1.1.2
gunicorn==20.1.0
h11==0.12.0
httpcore==0.15.0
httplib2==0.20.4
httpx==0.23.0
idna==3.2
importlib-metadata==4.11.4
inflection==0.5.1
itsdangerous==1.1.0
jeepney==0.7.1
Jinja2==2.11.3
joblib==1.1.0
jsonschema==4.6.0
keyring==9.3.1
keyrings.alt==3.2.0
kiwisolver==1.3.2
lazy-object-proxy==1.7.1
lockfile==0.12.2
Mako==1.2.0
Markdown==3.3.7
MarkupSafe==2.0.1
marshmallow==3.16.0
marshmallow-enum==1.5.1
marshmallow-oneofschema==3.0.1
marshmallow-sqlalchemy==0.26.1
matplotlib==3.5.0
nltk==3.7
numpy==1.21.2
oauthlib==3.2.0
packaging==21.3
pandas==1.3.3
pathspec==0.9.0
pendulum==2.1.2
Pillow==8.4.0
plotly==5.8.2
pluggy==1.0.0
prison==0.2.1
protobuf==3.19.1
psutil==5.9.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.12.0
PyJWT==1.7.1
pyparsing==3.0.6
pyrsistent==0.18.1
python-daemon==2.3.0
python-dateutil==2.8.2
python-dotenv==0.20.0
python-nvd3==0.15.0
python-slugify==6.1.2
python3-openid==3.2.0
pytz==2022.1
pytzdata==2020.1
PyYAML==6.0
regex==2022.4.24
requests==2.27.1
requests-oauthlib==1.3.1
rfc3986==1.5.0
rich==12.4.4
rsa==4.8
scipy==1.7.3
seaborn==0.11.2
SecretStorage==3.3.1
setproctitle==1.2.3
setuptools-scm==6.3.2
six==1.16.0
sniffio==1.2.0
soupsieve==2.3.1
SQLAlchemy==1.4.9
SQLAlchemy-JSONField==1.0.0
SQLAlchemy-Utils==0.38.2
swagger-ui-bundle==0.0.9
tabulate==0.8.9
tenacity==8.0.1
termcolor==1.1.0
text-unidecode==1.3
textblob==0.17.1
tomli==1.2.2
tqdm==4.64.0
typing-extensions==4.2.0
tzlocal==2.0.0
unicodecsv==0.14.1
uritemplate==4.1.1
urllib3==1.26.6
vaderSentiment==3.3.2
webassets==2.0
Werkzeug==1.0.1
wordcloud==1.8.1
wrapt==1.14.1
WTForms==2.3.3
zipp==3.8.0
